requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.3.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper==20231117
packaging==24.0
PyJWT==2.8.0
python-dotenv==1.0.1
regex==2024.5.10
requests==2.31.0
six==1.16.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.4
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
whisper==1.1.10
yarl==1.9.4
zipp==3.18.1
